Moved portal and added key to search-interactions
