+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/portal/index.tsx
