+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/presentation/dropdown/dropdown.tsx
@@ -14,9 +14,9 @@
  **/
 import * as React from 'react'
 import { hot } from 'react-hot-loader'
-import Portal from '../portal'
 import styled from 'styled-components'
 import { ThemeInterface } from '../../styles/styled-components'
+import Portal from '../../portal'
 import { Dropshadow } from '../../styles/mixins'
 import ChangeBackground from '../change-background'
 import ButtonBehavior from '../button-behavior'
